requirements.txt
Flask==2.2.3
flask_sqlalchemy==3.0.3
python_bcrypt==0.3.2
